Split `ExhaustArray` to a module and rename `core_primitive` to `primitive`.
The previous organization did not really make sense except as conflating
“types that are unconditionally available” with “types in `core`”.
